+++ pai-dsw-20220101/alibabacloud_pai_dsw20220101/models/_create_instance_request.py
@@ -2,7 +2,7 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
-from typing import List, Dict, Any
+from typing import Dict, List, Any
 
 from alibabacloud_pai_dsw20220101 import models as main_models
 from darabonba.model import DaraModel
@@ -10,12 +10,14 @@ from darabonba.model import DaraModel
 class CreateInstanceRequest(DaraModel):
     def __init__(
         self,
+        access_restriction_rules: Dict[str, str] = None,
         accessibility: str = None,
         affinity: main_models.CreateInstanceRequestAffinity = None,
         assign_node_spec: main_models.CreateInstanceRequestAssignNodeSpec = None,
         cloud_disks: List[main_models.CreateInstanceRequestCloudDisks] = None,
         credential_config: main_models.CredentialConfig = None,
         datasets: List[main_models.CreateInstanceRequestDatasets] = None,
+        docker_config: main_models.CreateInstanceRequestDockerConfig = None,
         driver: str = None,
         dynamic_mount: main_models.DynamicMount = None,
         ecs_spec: str = None,
@@ -38,6 +40,7 @@ class CreateInstanceRequest(DaraModel):
         workspace_id: str = None,
         workspace_source: str = None,
     ):
+        self.access_restriction_rules = access_restriction_rules
         # The instance accessibility.
         # 
         # Valid values:
@@ -54,6 +57,7 @@ class CreateInstanceRequest(DaraModel):
         self.credential_config = credential_config
         # The datasets.
         self.datasets = datasets
+        self.docker_config = docker_config
         # The NVIDIA driver configuration.
         self.driver = driver
         # The dynamic mount configuration.
@@ -126,6 +130,8 @@ class CreateInstanceRequest(DaraModel):
             for v1 in self.datasets:
                  if v1:
                     v1.validate()
+        if self.docker_config:
+            self.docker_config.validate()
         if self.dynamic_mount:
             self.dynamic_mount.validate()
         if self.labels:
@@ -150,6 +156,9 @@ class CreateInstanceRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.access_restriction_rules is not None:
+            result['AccessRestrictionRules'] = self.access_restriction_rules
+
         if self.accessibility is not None:
             result['Accessibility'] = self.accessibility
 
@@ -171,6 +180,9 @@ class CreateInstanceRequest(DaraModel):
         if self.datasets is not None:
             for k1 in self.datasets:
                 result['Datasets'].append(k1.to_map() if k1 else None)
+
+        if self.docker_config is not None:
+            result['DockerConfig'] = self.docker_config.to_map()
 
         if self.driver is not None:
             result['Driver'] = self.driver
@@ -243,6 +255,9 @@ class CreateInstanceRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AccessRestrictionRules') is not None:
+            self.access_restriction_rules = m.get('AccessRestrictionRules')
+
         if m.get('Accessibility') is not None:
             self.accessibility = m.get('Accessibility')
 
@@ -269,6 +284,10 @@ class CreateInstanceRequest(DaraModel):
             for k1 in m.get('Datasets'):
                 temp_model = main_models.CreateInstanceRequestDatasets()
                 self.datasets.append(temp_model.from_map(k1))
+
+        if m.get('DockerConfig') is not None:
+            temp_model = main_models.CreateInstanceRequestDockerConfig()
+            self.docker_config = temp_model.from_map(m.get('DockerConfig'))
 
         if m.get('Driver') is not None:
             self.driver = m.get('Driver')
@@ -692,6 +711,49 @@ class CreateInstanceRequestLabels(DaraModel):
 
         if m.get('Value') is not None:
             self.value = m.get('Value')
+
+        return self
+
+class CreateInstanceRequestDockerConfig(DaraModel):
+    def __init__(
+        self,
+        containers_limit: int = None,
+        enable: bool = None,
+        mount_access_config_path: str = None,
+    ):
+        self.containers_limit = containers_limit
+        self.enable = enable
+        self.mount_access_config_path = mount_access_config_path
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.containers_limit is not None:
+            result['ContainersLimit'] = self.containers_limit
+
+        if self.enable is not None:
+            result['Enable'] = self.enable
+
+        if self.mount_access_config_path is not None:
+            result['MountAccessConfigPath'] = self.mount_access_config_path
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ContainersLimit') is not None:
+            self.containers_limit = m.get('ContainersLimit')
+
+        if m.get('Enable') is not None:
+            self.enable = m.get('Enable')
+
+        if m.get('MountAccessConfigPath') is not None:
+            self.mount_access_config_path = m.get('MountAccessConfigPath')
 
         return self